Issue #30: Explicitly static cast to remove warning.

+++ reg-lib/cpu/_reg_ssd.cpp
@@ -644,9 +644,9 @@ void GetDiscretisedValueSSD_core3D(nifti_image *controlPointGridImage,
 
                DTYPE warpedValue;
                int paddedImageVox[3] = {
-                  imageVox[0]+warPaddedOffset[0],
-                  imageVox[1]+warPaddedOffset[1],
-                  imageVox[2]+warPaddedOffset[2]
+                  static_cast<int>(imageVox[0]+warPaddedOffset[0]),
+                  static_cast<int>(imageVox[1]+warPaddedOffset[1]),
+                  static_cast<int>(imageVox[2]+warPaddedOffset[2])
                };
                int cc;
                double currentSum;
